requirements.txt
Django>=4.2,<6
psycopg2-binary>=2.9
gunicorn>=21.0
python-decouple>=3.8
Pillow>=10.0
whitenoise>=6.0
